Extract seed reconciliation logic into dedicated function
Also, move related logic from `pkg/operation/seed` to the controller package so that it's not unnecessarily scattered and distributed.
