+++ pkg/operation/seed/types.go
@@ -20,7 +20,6 @@ import (
 	"sync/atomic"
 
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
-	"github.com/gardener/gardener/pkg/operation/botanist/component"
 )
 
 // Builder is an object that builds Seed objects.
@@ -34,11 +33,4 @@ type Seed struct {
 	infoMutex sync.Mutex
 
 	LoadBalancerServiceAnnotations map[string]string
-
-	components *Components
-}
-
-// Components contains different components deployed in the Seed cluster.
-type Components struct {
-	dnsRecord component.DeployMigrateWaiter
 }
